+++ src/components/GearIcon.tsx
@@ -1,0 +1,13 @@
+import './GearIcon.css';
+
+const GearIcon = () => {
+  return (
+    <img 
+      src="/images/gear.png" 
+      alt="Gear Icon"
+      className="gear-spin w-8 h-8"
+    />
+  );
+};
+
+export default GearIcon;